Removed all warnings, mostly type-casting stuff with floats and doubles. Fixed the "getDebugMode" function returning a bool rather than an int. Switched over the last few menu attributes to floats. Overloaded object "setAngle" and "setPosition" functions to accept "vector3d" objects. Fixed particle billboarding, re-enabled z-rotation, and switched z-rotation over to floats. Made particle randomization float-based. Added error messages when shader files can't be located.  Added "get" functions for camera and object rotation order. Added a function that returns the inverse order of a sequence of axes. Split the random range function into one that returns integers and one that returns floats. Added an orbit function and an almost-correctly-working "angleTo" function. Overloaded world "setCamAngle" and "setCamPosition" functions to accept "vector3d" objects.

+++ TDML/config.cpp
@@ -36,7 +36,7 @@ namespace TDML
 			configd[configd.size()-1].append(word);
 		}
 		multisample = atof(configd[0].c_str())!=0;
-		fov = atof(configd[1].c_str());
+		fov = (float) atof(configd[1].c_str());
 		Shaders.setShaderPath(configd[2].c_str());
 	}
